postgres/postgres.go
package postgres

import (
	"errors"
	"strings"

	"github.com/mstgnz/sqlporter"
)

type PostgreSQL struct {
	schema *sqlporter.Schema
}

// NewPostgreSQL creates a new PostgreSQL parser instance
func NewPostgreSQL() *PostgreSQL {
	return &PostgreSQL{
		schema: &sqlporter.Schema{},
	}
}

// Parse parses PostgreSQL dump content
func (p *PostgreSQL) Parse(content string) (*sqlporter.Schema, error) {
	if content == "" {
		return nil, errors.New("empty content")
	}

	// TODO: PostgreSQL dump parsing operations will be implemented here
	return p.schema, nil
}

// Generate generates PostgreSQL dump from schema
func (p *PostgreSQL) Generate(schema *sqlporter.Schema) (string, error) {
	if schema == nil {
		return "", errors.New("empty schema")
	}

	var result strings.Builder

	// TODO: PostgreSQL dump generation operations will be implemented here

	return result.String(), nil
}
